backend/requirements.txt
Flask
flask-cors
requests
opencv-python
nltk
google-cloud-vision